TOKIO_VD/requirements.txt
altair==5.0.1
streamlit==1.32.0
python-dotenv==0.21.1
boto3==1.26.60
requests==2.31.0
typing==3.7.4.3
pandas==1.5.3
pillow==10.3.0
openpyxl==3.1.2
